+++ lib/altamira/helpers/all.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'altamira/helpers/json_request'
+require 'altamira/helpers/config_reader'