+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Improved staging file lookup
   - Signals with negative gains now swap their physical minimum and physical
     maximum when autoscaling is off
+  - Sleep stages for hypnograms are now loaded directly from XML Annotation
+    files
 - **Gem Changes**
   - Updated to Ruby 2.4.0
   - Updated to edfize 0.4.0